Avoid empty folders when the internet speed is unstable
When the internet speed is unstable, the tool may just create a folder and
then interrupt. When I run the tool again, it will skip some images not fully
downloaded because the tool think having a folder is equivalent to having an
image. So I believe that we should not use the folder to determine if the
image is downloaded.